+++ spec/Component/AbstractComponentSpec.php
@@ -36,10 +36,10 @@ class AbstractComponentSpec extends ObjectBehavior
     function let(EngineInterface $templateEngine)
     {
         $this->modelData = [
-            'type' => 'test',
-            'headline' => serialize(['unit' => 'h1', 'value' => 'test']),
-            'id'   => 1,
-            'cssID' => serialize(['', '']),
+            'type'      => 'test',
+            'headline'  => serialize(['unit' => 'h1', 'value' => 'test']),
+            'id'        => 1,
+            'cssID'     => serialize(['', '']),
             'customTpl' => 'custom_tpl'
         ];